+++ drift_dev/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Support `MAPPED BY` for individual columns in queries or in views defined with SQL.
 - Consistently interpret `CAST (x AS DATETIME)` and `CAST(x AS TEXT)` in drift files.
 - Support a `CAST` to an enum type in drift types.
+- Support two different queries using `LIST()` columns having the same result class name.
 - Fix table classes not extending defining Dart classes with modular generation.
 - Fix `@UseDataClass` with `extending` not working with modular generation.